+++ sample_frontend_container/sample_remix_react/app/entry.client.tsx
@@ -8,6 +8,7 @@ import { RemixBrowser } from '@remix-run/react';
 import { startTransition, StrictMode } from 'react';
 import { hydrateRoot } from 'react-dom/client';
 
+// モックAPI用の設定
 if (process.env.NODE_ENV === 'development') {
   import('./mocks/browser').then(({ worker }) => {
     worker.start();
