+++ sample_frontend_container/sample_remix_react/app/mocks/browser.ts
@@ -1,4 +1,5 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers';
+// NOTE: ↓このimport文の書き方で./handlers/index.tsを自動で読み込んでhandlersを見つけてくれる。
+import { handlers } from './handlers'; 
 
 export const worker = setupWorker(...handlers);
