+++ sample_frontend_container/sample_remix_react/app/mocks/handlers/index.ts
@@ -1,0 +1,9 @@
+import { loginHandler } from './loginHandler';
+// 他のハンドラーもここでインポート
+// import { exampleHandler } from './exampleHandler';
+
+export const handlers = [
+  loginHandler,
+  // 他のハンドラーもここに追加
+  // exampleHandler,
+];
